fix: clean only if requested

+++ builder/macos.py
@@ -41,8 +41,10 @@ class MacBuilder(Builder):
 	def compile(self):
 		cmake_build_dir = os.path.join(self.dir_source, "blender-cmake-build")
 		if os.path.exists(cmake_build_dir):
-			utils.remove_directory(cmake_build_dir)
-		os.makedirs(cmake_build_dir)
+			if self.build_clean:
+				utils.remove_directory(cmake_build_dir)
+		else:
+			os.makedirs(cmake_build_dir)
 		os.chdir(cmake_build_dir)
 
 		cmake = ['cmake']
